% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in.R
\name{sign_in}
\alias{sign_in}
\title{Sign In to Tableau Server}
\usage{
sign_in(token_name, token_secret, server_url, site_url = "", api_version)
}
\arguments{
\item{token_name}{The name of the Tableau token}

\item{token_secret}{The Tableau personal access token}

\item{server_url}{Tableau Server URL}

\item{site_url}{The Site URL to log in (if left blank it logs into the default site)}

\item{api_version}{The Tableau REST API version}
}
\value{
None.
}
\description{
Set the credentials and authenticate for a selected Tableau Server. The sign in function sets a secret access token
that can be used for all following API requests. This secret access token is valid for 240 minutes and has to be
refreshed after this time period.
}
\examples{
\dontrun{
sign_in(
  token_name = "my_token",
  personal_access_token = "EE2D2915A4",
  server_url = "https://my-tableau-server.com",
  site_url = "my_site",
  api_version = 3.9)
}
}
