% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_out.R
\name{sign_out}
\alias{sign_out}
\title{Sign Out from Tableau Server}
\usage{
sign_out(
  server_url = Sys.getenv("TABELAU_SERVER_URL"),
  api_version = Sys.getenv("TABLEAU_API_VERSION"),
  api_token = Sys.getenv("TABLEAU_API_TOKEN")
)
}
\arguments{
\item{server_url}{Tableau Server URL}

\item{api_version}{The Tableau REST API version}

\item{api_token}{The API Access Token that was retrieved when signin in to Tableau Server}
}
\value{
None.
}
\description{
Unsets the credentials signs out from a selected Tableau Server.
}
\examples{
\dontrun{
sign_out(
  server_url = Sys.getenv("TABELAU_SERVER_URL"),
  api_version = Sys.getenv("TABLEAU_API_VERSION"),
  api_token = Sys.getenv("TABLEAU_API_TOKEN"))
}
}
