% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_to_site.R
\name{switch_to_site}
\alias{switch_to_site}
\title{Switch site}
\usage{
switch_to_site(
  server_url = Sys.getenv("TABLEAU_SERVER_URL"),
  api_version = Sys.getenv("TABLEAU_API_VERSION"),
  api_token = Sys.getenv("TABLEAU_API_TOKEN"),
  site_url
)
}
\arguments{
\item{server_url}{Tableau Server URL}

\item{api_version}{The Tableau REST API version}

\item{api_token}{The API Access Token that was retrieved when signing in to Tableau Server}

\item{site_url}{The URL of the site to switch to}
}
\value{
None.
}
\description{
Can be used to switch between sites on Tableau Server. This method is currently not available for Tableau Online.
}
\examples{
\dontrun{
switch_to_site(
  server_url = Sys.getenv("TABLEAU_SERVER_URL"),
  api_version = Sys.getenv("TABLEAU_API_VERSION"),
  api_token = Sys.getenv("TABLEAU_API_TOKEN"),
  site_url = "MarketingSite")
}
}
